Gemini Chatbot/requirement.txt
Flask==2.3.2
google-generativeai==0.3.1